feat(gimbal): split homing for pearl/diamond
split homing routines into several files for different hardware as they
are going to be very different.
Homing with one axis won't be possible on diamond: it won't be looking
for the center anymore but a flat mirror.
