+++ main_board/include/app_config.h
@@ -67,7 +67,8 @@
 #define THREAD_PRIORITY_ALS   11
 #define THREAD_STACK_SIZE_ALS 2048
 
-#define THREAD_PRIORITY_MOTORS_INIT 12
+#define THREAD_STACK_SIZE_MOTORS_INIT 2048 // * 2 threads
+#define THREAD_PRIORITY_MOTORS_INIT   12
 
 #define THREAD_PRIORITY_BATTERY   8
 #define THREAD_STACK_SIZE_BATTERY 2048
